requirements.txt
numpy
opencv-python
tqdm
torch
mediapipe
scikit-learn
matplotlib
PyYAML
seaborn